requirements.txt
# Pinned for reproducibility
Django==6.0.1
numpy==2.4.1
Pillow==12.1.0
scikit-learn==1.8.0
tensorflow==2.20.0
h5py==3.15.1
joblib==1.4.2
python-dotenv==1.0.0
